+++ artiq/test/coredevice/test_rtio.py
@@ -12,6 +12,7 @@ from artiq.coredevice import exceptions
 from artiq.coredevice.comm_mgmt import CommMgmt
 from artiq.coredevice.comm_analyzer import (StoppedMessage, OutputMessage, InputMessage,
                                             decode_dump, get_analyzer_dump)
+from artiq.compiler.targets import CortexA9Target
 
 
 artiq_low_latency = os.getenv("ARTIQ_LOW_LATENCY")
@@ -230,17 +231,18 @@ class LoopbackGateTiming(EnvExperiment):
         # With the exact delay known, make sure tight gate timings work.
         # In the most common configuration, 24 mu == 24 ns == 3 coarse periods,
         # which should be plenty of slack.
+        # FIXME: ZC706 with NIST_QC2 needs 48ns - hw problem?
         delay_mu(10000)
 
         gate_start_mu = now_mu()
-        self.loop_in.gate_both_mu(24)
+        self.loop_in.gate_both_mu(48) # XXX
         gate_end_mu = now_mu()
 
         # gateware latency offset between gate and input
         lat_offset = 11*8
         out_mu = gate_start_mu - loop_delay_mu + lat_offset
         at_mu(out_mu)
-        self.loop_out.pulse_mu(24)
+        self.loop_out.pulse_mu(48) # XXX
 
         in_mu = self.loop_in.timestamp_mu(gate_end_mu)
         print("timings: ", gate_start_mu, in_mu - lat_offset, gate_end_mu)
@@ -460,11 +462,15 @@ class CoredeviceTest(ExperimentCase):
 
     def test_pulse_rate(self):
         """Minimum interval for sustained TTL output switching"""
-        self.execute(PulseRate)
+        exp = self.execute(PulseRate)
         rate = self.dataset_mgr.get("pulse_rate")
         print(rate)
         self.assertGreater(rate, 100*ns)
-        self.assertLess(rate, 480*ns)
+        if exp.core.target_cls == CortexA9Target:
+            # Crappy AXI PS/PL interface from Xilinx is slow.
+            self.assertLess(rate, 810*ns)
+        else:
+            self.assertLess(rate, 480*ns)
 
     def test_pulse_rate_ad9914_dds(self):
         """Minimum interval for sustained AD9914 DDS frequency switching"""
@@ -621,11 +627,13 @@ class _DMA(EnvExperiment):
         self.delta = now_mu() - start
 
     @kernel
-    def playback_many(self, n):
+    def playback_many(self, n, add_delay=False):
         handle = self.core_dma.get_handle(self.trace_name)
         self.core.break_realtime()
         t1 = self.core.get_rtio_counter_mu()
         for i in range(n):
+            if add_delay:
+                delay(2*us)
             self.core_dma.playback_handle(handle)
         t2 = self.core.get_rtio_counter_mu()
         self.set_dataset("dma_playback_time", self.core.mu_to_seconds(t2 - t1))
@@ -718,13 +726,18 @@ class DMATest(ExperimentCase):
             self.device_mgr.get_desc("ad9914dds0")
         except KeyError:
             raise unittest.SkipTest("skipped on Kasli for now")
+
         exp = self.create(_DMA)
+        is_zynq = exp.core.target_cls == CortexA9Target
         count = 20000
         exp.record_many(40)
-        exp.playback_many(count)
+        exp.playback_many(count, is_zynq)
         dt = self.dataset_mgr.get("dma_playback_time")
         print("dt={}, dt/count={}".format(dt, dt/count))
-        self.assertLess(dt/count, 4.5*us)
+        if is_zynq:
+            self.assertLess(dt/count, 6.2*us)
+        else:
+            self.assertLess(dt/count, 4.5*us)
 
     def test_dma_underflow(self):
         exp = self.create(_DMA)